refactor: Make MetaDB Datastore only
Remove the intermediate MetaDB interface and move methods in
the datastore.Driver struct to the new MetaDB struct.
We will focus on the GCP implementation and remove the redundant code.
