+++ internal/app/server/open_saves.go
@@ -25,7 +25,6 @@ import (
 	"github.com/googleforgames/open-saves/internal/pkg/blob"
 	"github.com/googleforgames/open-saves/internal/pkg/cache"
 	"github.com/googleforgames/open-saves/internal/pkg/metadb"
-	"github.com/googleforgames/open-saves/internal/pkg/metadb/datastore"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -63,13 +62,9 @@ func newOpenSavesServer(ctx context.Context, cloud, project, bucket, cacheAddr s
 		if err != nil {
 			return nil, err
 		}
-		datastore, err := datastore.NewDriver(ctx, project)
+		metadb, err := metadb.NewMetaDB(ctx, project)
 		if err != nil {
-			return nil, err
-		}
-		metadb := metadb.NewMetaDB(datastore)
-		if err := metadb.Connect(ctx); err != nil {
-			log.Fatalf("Failed to connect to the metadata server: %v", err)
+			log.Fatalf("Failed to create a MetaDB instance: %v", err)
 			return nil, err
 		}
 		redis := cache.NewRedis(cacheAddr)
