+++ internal/pkg/metadb/error.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package datastore
+package metadb
 
 import (
 	ds "cloud.google.com/go/datastore"
